backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pydantic==2.12.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
openai==1.99.9
apscheduler==3.11.2
python-multipart==0.0.22
python-dotenv==1.2.1
httpx==0.28.1
bcrypt==4.1.3
PyJWT==2.11.0
